refactor(helpers): split out functions into files
This will help with maintainability and with tree-shaking when packaging separately for cjs.
