+++ src/Support/array/index.ts
@@ -1,0 +1,1 @@
+export { default as wrap } from './wrap';
