+++ src/Support/function/index.ts
@@ -1,0 +1,5 @@
+export { default as dataGet } from './dataGet';
+export { default as retry } from './retry';
+export { default as transformKeys } from './transformKeys';
+export { default as isUserLandClass } from './isUserLandClass';
+export { default as isObjectLiteral } from './isObjectLiteral';
