+++ src/Support/initialiser/collect.ts
@@ -1,0 +1,12 @@
+import Collection from '../Collection';
+
+/**
+ * Create a collection from the array.
+ *
+ * @param {any} items
+ *
+ * @return {Collection}
+ */
+export default function collect<T>(items?: T | T[]): Collection<T> {
+    return new Collection(items);
+}
